Rename ribbon to pyqtribbon
