+++ .github/workflows/pytest.yaml
@@ -5,7 +5,7 @@ on:
     branches:
     - main
     paths:
-    - 'ribbon/**'
+    - 'pyqtribbon/**'
   # pull_request event is required only for autolabeler
   pull_request:
     # Only following types are handled by the action, but one can default to all as well
