+++ docs/source/_screenshots/panel.py
@@ -4,8 +4,8 @@ from qtpy.QtWidgets import QApplication, QToolButton, QMenu, QLabel, QLineEdit
 from qtpy.QtGui import QIcon
 from qtpy.QtCore import Qt
 
-from ribbon import RibbonBar
-from ribbon.screenshotwindow import RibbonScreenShotWindow
+from pyqtribbon import RibbonBar
+from pyqtribbon.screenshotwindow import RibbonScreenShotWindow
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)
