+++ docs/source/user.rst
@@ -20,7 +20,7 @@ you can use the `setMenuBar` method of :py:class:`~PyQt5.QtWidgets.QMainWindow` 
 .. code-block:: python
 
     ...
-    from ribbon import RibbonBar
+    from pyqtribbon import RibbonBar
 
     window = QtWidgets.QMainWindow()
     ribbon = RibbonBar()
