+++ pyqtribbon/Qt.py
