+++ pyqtribbon/__init__.py
