+++ pyqtribbon/category.py
