+++ pyqtribbon/gallery.py
