+++ pyqtribbon/panel.py
