+++ pyqtribbon/ribbonbar.py
