+++ pyqtribbon/separator.py
