+++ pyqtribbon/tabbar.py
