+++ pyqtribbon/titlewidget.py
