+++ pyqtribbon/typehints.py
