+++ pyqtribbon/utils.py
